+++ src/Zicht/Bundle/KeyValueBundle/DependencyInjection/ZichtKeyValueExtension.php
@@ -5,7 +5,6 @@
 
 namespace Zicht\Bundle\KeyValueBundle\DependencyInjection;
 
-use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
@@ -35,38 +34,13 @@ class ZichtKeyValueExtension extends Extension
 
         $rootDir = $container->getParameter('kernel.root_dir');
         $container->getDefinition('zicht_bundle_key_value.key_value_storage_manager')
-            ->replaceArgument(1, realpath(sprintf('%s/%s', $rootDir, '../web')))
-            ->replaceArgument(2, $this->checkDirectory(sprintf('%s/%s', $rootDir, '../web/media/key_value_storage')));
+            ->replaceArgument(1, sprintf('%s/%s', $rootDir, '../web'))
+            ->replaceArgument(2, sprintf('%s/%s', $rootDir, '../web/media/key_value_storage'));
 
         $container->getDefinition('zicht_bundle_key_value.form_type.locale_dependent_data_type')
             ->replaceArgument(0, $config['locales']);
 
         $container->getDefinition('zicht_bundle_key_value.event_listener.localization_listener')
             ->replaceArgument(0, array_map(function ($value) { return $value['locale']; }, $config['locales']));
-    }
-
-    /**
-     * Check, at compile time, if the directories are available and writable
-     *
-     * @param string $directory
-     * @return string
-     * @throws InvalidConfigurationException
-     */
-    private function checkDirectory($directory)
-    {
-        if (!is_dir($directory)) {
-            @mkdir($directory);
-        }
-
-        $realDirectory = realpath($directory);
-        if (false === $realDirectory) {
-            throw new InvalidConfigurationException(sprintf('Directory does not exist [%s]', $directory));
-        }
-
-        if (!is_writeable($realDirectory)) {
-            throw new InvalidConfigurationException(sprintf('Directory is not writable [%s]', $realDirectory));
-        }
-
-        return $realDirectory;
     }
 }